+++ src/Misc/Hooks.BugFixes.cpp
@@ -1175,7 +1175,15 @@ DEFINE_HOOK(0x7185DA, TeleportLocomotionClass_MakeRoom_DestFix, 0x6)
 
 	GET(CellStruct*, pCellAt, EAX);
 
-	return *pCellAt == CellStruct::Empty ? ReturnTrue : 0;
+	if (*pCellAt == CellStruct::Empty)
+	{
+		GET(const LocomotionClass* const, pLoco, EBP);
+		// cannot find location ? dont move
+		pLoco->LinkedTo->ChronoDestCoords = pLoco->LinkedTo->Location;
+		return ReturnTrue;
+	}
+
+	return 0;
 }
 
 #pragma region TeleportLocomotionOccupationFix
@@ -1361,6 +1369,21 @@ DEFINE_HOOK(0x6F4BB3, TechnoClass_ReceiveCommand_RequestUntether, 0x7)
 }
 
 #pragma endregion
+
+// Fix the bug that techno in attack move will move to target if it cannot attack it
+DEFINE_HOOK(0x4D77BD, FootClass_ObjectClickedAction_NoMove, 0x6)
+{
+	enum { Attack = 0x4D769F };
+
+	GET(ObjectClass*, pTarget, EBX);
+	const auto pTargetTechno = abstract_cast<TechnoClass*>(pTarget);
+
+	if (!pTargetTechno)
+		return Attack;
+
+	GET(FootClass*, pThis, ESI);
+	return pThis->Owner->IsAlliedWith(pTargetTechno->Owner) ? 0 : Attack;
+}
 
 #pragma region JumpjetShadowPointFix